+++ includes/globals.php
@@ -5,7 +5,7 @@
  */
 
 // Current CMS version
-const VERSION = '1.7.2';
+const VERSION = '1.7.3';
 
 /**
  * Display copyright on the admin dashboard.