+++ includes/schema.php
@@ -97,6 +97,7 @@ function dbSchema() {
 	$tables['user_roles'] = "CREATE TABLE user_roles (
 		id bigint(20) unsigned PRIMARY KEY auto_increment,
 		name varchar(255) NOT NULL,
+		_default varchar(5) NOT NULL,
 		KEY name (name)
 	);";